@@ -14,7 +14,8 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 """Quantum state object."""
-from typing import List, Optional, Tuple
+from __future__ import annotations
+from typing import Optional
 
 import numpy as np
 
@@ -25,7 +26,7 @@ from toqito.matrix_props import is_density
 class State:
     """A `State` object representing a quantum state."""
 
-    def __init__(self, state: np.ndarray, dims: List[int]) -> None:
+    def __init__(self, state: np.ndarray, dims: list[int]) -> None:
         """Initializes a quantum state.
 
         Args:
@@ -36,7 +37,7 @@ class State:
         self._dims = self._prepare_dims(dims)
         self._systems = list(range(1, len(self._dims) + 1))
 
-    def __eq__(self, other: "State") -> bool:
+    def __eq__(self, other: State) -> bool:
         if isinstance(other, State):
             return (
                 np.allclose(self.value, other.value)
@@ -67,23 +68,23 @@ class State:
         return self.__str__()
 
     @property
-    def shape(self) -> Tuple[int, int]:
+    def shape(self) -> tuple[int, int]:
         return self._state.shape
 
     @property
-    def dims(self) -> List[int]:
+    def dims(self) -> list[int]:
         return self._dims
 
     @property
-    def systems(self) -> List[int]:
+    def systems(self) -> list[int]:
         return self._systems
 
     @property
-    def alice_systems(self) -> List[int]:
+    def alice_systems(self) -> list[int]:
         return [i for i in self._systems if i % 2 != 0]
 
     @property
-    def bob_systems(self) -> List[int]:
+    def bob_systems(self) -> list[int]:
         return [i for i in self._systems if i % 2 == 0]
 
     @property
@@ -117,7 +118,7 @@ class State:
 
         return state
 
-    def _prepare_dims(self, dims: List[int]) -> Optional[List[int]]:
+    def _prepare_dims(self, dims: list[int]) -> Optional[list[int]]:
         """Returns the validated list of dimensions to be used for the quantum state.
 
         Args:
@@ -134,7 +135,7 @@ class State:
             )
         return dims
 
-    def kron(self, r_state: "State") -> "State":
+    def kron(self, r_state: State) -> State:
         """Performs the Kronecker (tensor) product between two states.
 
         Args:
@@ -144,7 +145,7 @@ class State:
         new_dims = self._dims + r_state.dims
         return State(new_state, new_dims)
 
-    def swap(self, sub_sys_swap: List[int]) -> None:
+    def swap(self, sub_sys_swap: list[int]) -> None:
         """Performs a swap between two subsystems of the state.
 
         Args:
